requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
onnxruntime>=1.17.0
